Http-pyyntö ListPricesFramentista omaan HttpService-luokkaan
